+++ lib/source/pl/lib/std/file.cpp
@@ -33,7 +33,7 @@ namespace pl::lib::libstd::file {
             /* open(path, mode) */
             runtime.addDangerousFunction(nsStdFile, "open", FunctionParameterCount::exactly(2), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
                 const auto path     = Token::literalToString(params[0], false);
-                const auto modeEnum = Token::literalToUnsigned(params[1]);
+                const auto modeEnum = (u32)Token::literalToUnsigned(params[1]);
 
                 hlp::fs::File::Mode mode;
                 switch (modeEnum) {
@@ -63,7 +63,7 @@ namespace pl::lib::libstd::file {
 
             /* close(file) */
             runtime.addDangerousFunction(nsStdFile, "close", FunctionParameterCount::exactly(1), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
@@ -75,8 +75,8 @@ namespace pl::lib::libstd::file {
 
             /* read(file, size) */
             runtime.addDangerousFunction(nsStdFile, "read", FunctionParameterCount::exactly(2), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
-                const auto size = Token::literalToUnsigned(params[1]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
+                const auto size = (u64)Token::literalToUnsigned(params[1]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
@@ -88,7 +88,7 @@ namespace pl::lib::libstd::file {
 
             /* write(file, data) */
             runtime.addDangerousFunction(nsStdFile, "write", FunctionParameterCount::exactly(2), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
                 const auto data = Token::literalToString(params[1], true);
 
                 if (!openFiles.contains(file))
@@ -101,8 +101,8 @@ namespace pl::lib::libstd::file {
 
             /* seek(file, offset) */
             runtime.addDangerousFunction(nsStdFile, "seek", FunctionParameterCount::exactly(2), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file   = Token::literalToUnsigned(params[0]);
-                const auto offset = Token::literalToUnsigned(params[1]);
+                const auto file   = (u32)Token::literalToUnsigned(params[0]);
+                const auto offset = (u64)Token::literalToUnsigned(params[1]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
@@ -114,7 +114,7 @@ namespace pl::lib::libstd::file {
 
             /* size(file) */
             runtime.addDangerousFunction(nsStdFile, "size", FunctionParameterCount::exactly(1), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
@@ -124,8 +124,8 @@ namespace pl::lib::libstd::file {
 
             /* resize(file, size) */
             runtime.addDangerousFunction(nsStdFile, "resize", FunctionParameterCount::exactly(2), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
-                const auto size = Token::literalToUnsigned(params[1]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
+                const auto size = (u64)Token::literalToUnsigned(params[1]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
@@ -137,7 +137,7 @@ namespace pl::lib::libstd::file {
 
             /* flush(file) */
             runtime.addDangerousFunction(nsStdFile, "flush", FunctionParameterCount::exactly(1), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
@@ -149,7 +149,7 @@ namespace pl::lib::libstd::file {
 
             /* remove(file) */
             runtime.addDangerousFunction(nsStdFile, "remove", FunctionParameterCount::exactly(1), [](Evaluator *, auto params) -> std::optional<Token::Literal> {
-                const auto file = Token::literalToUnsigned(params[0]);
+                const auto file = (u32)Token::literalToUnsigned(params[0]);
 
                 if (!openFiles.contains(file))
                     throwInvalidFileError();
